requirements.txt
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cors==1.0.1
DateTime==4.3
distlib==0.3.4
et-xmlfile==1.1.0
filelock==3.4.2
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
msgpack==1.0.3
Naked==0.1.31
numpy==1.22.1
oauth2client==4.1.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pep8==1.7.1
platformdirs==2.4.1
proto-plus==1.19.9
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.13.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jwt==3.3.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rsa==4.8
shellescape==3.8.1
six==1.16.0
sodapy==2.1.0
SQLAlchemy==1.4.31
sseclient==0.0.27
tldextract==3.1.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
